requirements.txt
Django>=4.2
djangorestframework>=3.14
gunicorn>=21.2
psycopg2-binary>=2.9
django-redis>=5.2.0,<6.0